Loaded Posts
Query our new backend for all posts using the index route.

Use the angular.copy() method to create a deep copy of the returned
data. This ensures that the $scope.posts variable in MainCtrl will also
be updated, ensuring the new values are reflect in our view.

Using the Angular $http service to query our posts route. The success()
function allows us to bind function that will be executed when the
request returns. Because our route will return a list of posts, all we
need to do is copy that list to the client side posts object. Notice
that we're using the angular.copy() function to do this as it will make
our UI update properly.

Now we need to call this function at an appropriate time to load the
data. We do this by adding a property called resolve to our home state.